+++ yourtour/yt-web/src/main/java/com/yt/vo/route/RouteScheduleVO.java
@@ -8,7 +8,7 @@ import com.yt.core.utils.DateUtils;
 import com.yt.vo.BaseVO;
 
 public class RouteScheduleVO extends BaseVO {
-	private int index;
+	private long index;
 	private long date;
 	private int  days;
 	private String placeIds;
@@ -55,11 +55,11 @@ public class RouteScheduleVO extends BaseVO {
 		super();
 	}
 
-	public int getIndex() {
+	public long getIndex() {
 		return index;
 	}
 
-	public void setIndex(int index) {
+	public void setIndex(long index) {
 		this.index = index;
 	}
